ecnmarker: add flag to build CO-RE program
Loading a BPF CO-RE program requires the Linux kernel to be built with
CONFIG_DEBUG_INFO_BTF=y. OpenWrt 22.03 does not have a build option for
this, so add a flag to build ecnmarker as a CO-RE program.

This makes it possible to build ecnmarker in OpenWrt buildroot as a non
CO-RE program, so we can use it on OpenWrt 22.03.

Leave the possibility to build as CO-RE program, as this drastically
speeds up development time.

  $ time (git clean -fdx; cmake .; make )

  real    0m1,278s
  user    0m0,598s
  sys     0m0,671s

  $ time make package/ecnmarker/{clean,compile} V=sc

  real    0m40,316s
  user    0m31,604s
  sys     0m11,222s

Signed-off-by: Stijn Tintel <stijn@linux-ipv6.be>