ecnmarker.uc: track state
If we want to reliably cleanup the tc filter on shutdown, we should
track the state. If we would cleanup based on the UCI config, and this
config has changed in between starting and stopping the service, we
would not clean up the correct tc filter(s). Track state to avoid this.

Signed-off-by: Stijn Tintel <stijn@linux-ipv6.be>