procd: restart service on interface changes
When doing ECN marking on interfaces that are dynamically created, like
wireless interfaces, it's possible the ecnmarker service will be started
before the interface actually exists. In this case, we'll end up not
marking anything.

Add procd netdev param and trigger to restart the service when one of
the interfaces configured in ecnmarker changes.

Signed-off-by: Stijn Tintel <stijn@linux-ipv6.be>