ecnmarker.uc: return on failure in attach_bpf
The stop_service function stops the ecnmarker service via ubus. This
does not happen immediately, so we still try to setup the tc filter even
if we failed to setup the tc qdisc. We might also add incorrect data in
the state.

Use return on failure in attach_bpf, right after calling the
stop_service function, to avoid running a command that will always fail,
or having incorrect data in the state.

Signed-off-by: Stijn Tintel <stijn@linux-ipv6.be>