ecnmarker.uc: improve logging in toggle_ecnmarker_bpf
If the BPF program is not loaded, the map is not available and the
command to toggle the value in the map will fail, but we're still
logging a message that we're enabling ECN-CE marking. We also end up
running toggle_ecnmarker_bpf every 1000ms, due to the command failure.

This could result in very confusing log messages:

  Tue Oct  4 14:58:07 2022 daemon.info ecnmarker: ECN-CE marking disabled, enabling for 5000 ms
  Tue Oct  4 14:58:08 2022 daemon.info ecnmarker: ECN-CE marking disabled, enabling for 5000 ms
  Tue Oct  4 14:58:09 2022 daemon.info ecnmarker: ECN-CE marking disabled, enabling for 5000 ms

Improve this by only logging the message if the command succeeded.

Signed-off-by: Stijn Tintel <stijn@linux-ipv6.be>