ecnmarker.uc: don't stop service if interface fails
Since we support multiple interfaces, we don't want to stop the service
if we fail to setup the tc qdisc and filter on one of them. We might
still be able to set them up on one of the next configured interfaces.

Instead, stop the service if all interfaces failed to setup.

Signed-off-by: Stijn Tintel <stijn@linux-ipv6.be>