procd: bail out if no interfaces enabled
When no interfaces are enabled in the UCI config, we start the service
anyway. This could lead to confusion. Add checks in the init script
that prints a message to the console and then bails out if no interfaces
are enabled. This should make it clear to the user they did not
configure the service properly.

Signed-off-by: Stijn Tintel <stijn@linux-ipv6.be>